Refactor and improve observations function
Previous update caused the `observations` function to return a 3-tuple of
`Any[]` because of an under-the-hood call to `observations_matrix`, a
bad function that I wrote when I was in a rush. This is no longer the
case, and I've added tests to verify correct return types to ensure we
don't see a regression.